+++ src/AngleSharp.Core.Tests/Library/ContextLoading.cs
@@ -65,7 +65,7 @@ namespace AngleSharp.Core.Tests.Library
             if (Helper.IsNetworkAvailable())
             {
                 var title = "PostUrlencodeNormal";
-                var address = "http://anglesharp.azurewebsites.net/PostUrlEncodeNormal";
+                var address = "https://anglesharp-tests.anglevisions.com/PostUrlEncodeNormal";
                 var config = Configuration.Default.WithDefaultLoader();
                 var document = await BrowsingContext.New(config).OpenAsync(address);
                 var h1 = document.QuerySelector("h1");
@@ -85,7 +85,7 @@ namespace AngleSharp.Core.Tests.Library
         {
             if (Helper.IsNetworkAvailable())
             {
-                var address = "http://anglesharp.azurewebsites.net/PostUrlEncodeNormal";
+                var address = "https://anglesharp-tests.anglevisions.com/PostUrlEncodeNormal";
                 var config = Configuration.Default.WithDefaultLoader();
                 var context = BrowsingContext.New(config);
                 var document = await context.OpenAsync(address);
@@ -120,7 +120,7 @@ namespace AngleSharp.Core.Tests.Library
         {
             if (Helper.IsNetworkAvailable())
             {
-                var address = "http://anglesharp.azurewebsites.net/";
+                var address = "https://anglesharp-tests.anglevisions.com/";
                 var config = Configuration.Default.WithDefaultLoader();
                 var context = BrowsingContext.New(config);
                 var document = await context.OpenAsync(address);
@@ -140,7 +140,7 @@ namespace AngleSharp.Core.Tests.Library
             if (Helper.IsNetworkAvailable())
             {
                 var title = "PostUrlencodeNormal";
-                var address = "http://anglesharp.azurewebsites.net/";
+                var address = "https://anglesharp-tests.anglevisions.com/";
                 var config = Configuration.Default.WithDefaultLoader();
                 var context = BrowsingContext.New(config);
                 var document = await context.OpenAsync(address);
@@ -184,7 +184,7 @@ namespace AngleSharp.Core.Tests.Library
         {
             if (Helper.IsNetworkAvailable())
             {
-                var address = "https://www.powerball.com";
+                var address = "https://anglesharp-tests.anglevisions.com/test-cases/powerball";
                 var config = Configuration.Default.WithLocaleBasedEncoding().WithPageRequester();
                 var context = BrowsingContext.New(config);
                 var document = await context.OpenAsync(address);
@@ -202,7 +202,7 @@ namespace AngleSharp.Core.Tests.Library
         {
             if (Helper.IsNetworkAvailable())
             {
-                var address = "http://kommersant.ru/rss-list";
+                var address = "https://anglesharp-tests.anglevisions.com/test-cases/kommersant";
                 var config = Configuration.Default.WithPageRequester();
                 var context = BrowsingContext.New(config);
                 var document = await context.OpenAsync(address);
@@ -217,7 +217,7 @@ namespace AngleSharp.Core.Tests.Library
         {
             if (Helper.IsNetworkAvailable())
             {
-                var address = "http://eurobelarus.info/";
+                var address = "https://anglesharp-tests.anglevisions.com/test-cases/eurobelarus";
                 var config = Configuration.Default.WithPageRequester();
                 var context = BrowsingContext.New(config);
                 var document = await context.OpenAsync(address);
@@ -235,7 +235,7 @@ namespace AngleSharp.Core.Tests.Library
             // * client is connected to VPN (at least with the VPN client of Windows 10).
             if (Helper.IsNetworkAvailable())
             {
-                var address = "http://anglesharp.azurewebsites.net/Chunked";
+                var address = "https://anglesharp-tests.anglevisions.com/Chunked";
                 var config = Configuration.Default.WithDefaultLoader();
                 var context = BrowsingContext.New(config);
                 var events = new EventReceiver<HtmlParseEvent>(handler => context.GetService<IHtmlParser>().Parsing += handler);
@@ -269,7 +269,7 @@ namespace AngleSharp.Core.Tests.Library
                 //However: The connection to taobao is usually very bad and the
                 //page takes ~10-30s (or longer!) to load. Replaced with another
                 //solution taken directly from the AngleSharp infrastructure.
-                var address = "http://anglesharp.azurewebsites.net/Page";
+                var address = "https://anglesharp-tests.anglevisions.com/Page";
                 var config = Configuration.Default.WithDefaultLoader(new LoaderOptions { IsResourceLoadingEnabled = true });
                 var context = BrowsingContext.New(config);
                 var document = await context.OpenAsync(address);
@@ -282,7 +282,7 @@ namespace AngleSharp.Core.Tests.Library
         {
             if (Helper.IsNetworkAvailable())
             {
-                var address = "http://www.europarl.europa.eu/sides/getDoc.do?type=COMPARL&reference=PE-583.901&format=PDF&language=EN&secondRef=01";
+                var address = "https://anglesharp-tests.anglevisions.com/test-cases/europarl";
                 var config = Configuration.Default.WithDefaultLoader();
                 var context = BrowsingContext.New(config);
                 var document = await context.OpenAsync(address);