+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.8.0
+
+Released on Saturday, September 5 2026
+
+- Improved AngleSharp's test website
+- Fixed script data escaped state potentially not bouncing back correctly
+- Added the `DomSameObject` annotation for the respective IDL members (#1314) @lahma
+- Added the `DomReturnType` annotation for methods returning a different IDL type (#1318)
+
 # 1.7.3
 
 Released on Thursday, September 3 2026